+++ examples/solid-ts/src/pages/menu.tsx
@@ -23,7 +23,7 @@ export default function Page() {
 
           <Portal>
             <div {...api().positionerProps}>
-              <ul class="menu-content" {...api().contentProps}>
+              <ul {...api().contentProps}>
                 <li {...api().getItemProps({ id: "edit" })}>Edit</li>
                 <li {...api().getItemProps({ id: "duplicate" })}>Duplicate</li>
                 <li {...api().getItemProps({ id: "delete" })}>Delete</li>