+++ examples/vue-ts/src/pages/dialog.tsx
@@ -32,7 +32,7 @@ export default defineComponent({
                 <div style={{ minHeight: "1200px", pointerEvents: "none" }} />
                 {parentDialog.isOpen && (
                   <Teleport to="body">
-                    <div class="dialog-backdrop" {...parentDialog.backdropProps} />
+                    <div {...parentDialog.backdropProps} />
                     <div {...parentDialog.containerProps} data-testid="container-1">
                       <div {...parentDialog.contentProps}>
                         <h2 {...parentDialog.titleProps}>Edit profile</h2>