+++ packages/machines/select/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @zag-js/select
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [[`5bd24f02`](https://github.com/chakra-ui/zag/commit/5bd24f02fcab355f7df8a2d5cea3b155155380f8)]:
+  - @zag-js/anatomy@0.1.2
+
+## 0.1.4
+
+### Patch Changes
+
+- [`cdef3fb0`](https://github.com/chakra-ui/zag/commit/cdef3fb08d61557bee50cdba63458406b731690e) Thanks
+  [@segunadebayo](https://github.com/segunadebayo)! - Fix package entrypoints
+
+## 0.1.3
+
+### Patch Changes
+
+- [`af4ab9bb`](https://github.com/chakra-ui/zag/commit/af4ab9bb7cd599c53e47ca7ed2ea90a4ff742499) Thanks
+  [@segunadebayo](https://github.com/segunadebayo)! - Update anatomy of components
+
+- [#445](https://github.com/chakra-ui/zag/pull/445)
+  [`3e242276`](https://github.com/chakra-ui/zag/commit/3e242276689016542e8a804329971bf493b8ba50) Thanks
+  [@visualjerk](https://github.com/visualjerk)! - Fix aria-labelledby on select trigger
+
+- Updated dependencies [[`9d936614`](https://github.com/chakra-ui/zag/commit/9d93661439f10a550c154e9f290905d32e8f509b),
+  [`9332f1ca`](https://github.com/chakra-ui/zag/commit/9332f1caf5122c16a3edb48e20664b04714d226c)]:
+  - @zag-js/core@0.2.3
+  - @zag-js/popper@0.2.1
+
 ## 0.1.2
 
 ### Patch Changes